Add files via upload
The Solar Analyst model itself is integrated with the workflow as a Python Script. The script executes an external instance of Solar Analyst in ESRI ArcGIS. Other operations such as inputs (parameters and raster DSM) and outputs (simple tabular output ) are also represented with labelled boxes. Lines connecting boxes capture flows of data or parameters between operations, left to right. Importantly, the workflow is executable, integrating both data and the operations upon data. The traffic lights associated with operations summarise whether the underlying data has been processed (green), is ready to be processed(yellow), or is not yet ready to be processed (red).