Add files via upload
Using a SWFMS helps make model validation easier to integrate. This workflow extends the minimal KNIME workflow for Solar Analyst in  with additional validation of the model output through comparison with a ground truth dataset.  The ground truth dataset included monthly average solar radiation potential kWh/m^2 of the study area for the last 30 years. In a particular case when the ground truth dataset is not available, average solar radiation over a specific period can be derived from satellite images for the area of interest https://power.larc.nasa.gov. The lower portion of the workflow extends the minimal KNIME workflow  by reading in and reformatting the ground truth dataset, calculating the differences between the model output and ground truth, and classifying those differences for ease of interpretation. 

The Solar Analyst output (record) is defined as an estimated solar radiation potential kWh/$m^2 for each raster cell in a DSM. Model outputs are classified based on their difference with the monthly average solar radiation potential value (AVG-RAD) derived from the ground truth data set. Model outputs with the maximum of 15\% relative difference (D) to the ground truth data set and within a range -15\% of AVG-RAD =< D =< 15\% of AVG-RAD are included in "expected" solar radiation potential category. Model outputs with a negative or positive difference value outside of the range are included in "low" and "high" solar radiation potential category, respectively. These three categories are specified on the difference value and do not represent nomenclature in any scientific domain. 
This simple validation method is based on the comparison between estimated and observed solar radiation values which enables inexperienced and other users an effective, automated and fast check for errors in modelling results.